refactor(bench): split payload generation helpers
