+++ docs/architecture/maintainability-baseline.txt
@@ -1,6 +1,5 @@
 # Maximum allowed production-line counts for current oversized Rust source files.
 # Format: <max_lines> <path>
-605 crates/surge-bench/src/payload.rs
 1013 crates/surge-bench/src/runner.rs
 813 crates/surge-cli/src/commands/install/mod.rs
 1860 crates/surge-cli/src/commands/install/remote.rs
